CLI: Rename eventEmitter to EventPromised
