+++ packages/frontity/src/utils/eventPromised.ts
